TTS/requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
bitsandbytes==0.49.2
boto3==1.42.59
botocore==1.42.59
brotli==1.2.0
cached_path==1.8.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coqpit-config==0.2.4
coqui-tts==0.26.2
coqui-tts-trainer==0.2.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
datasets==4.6.1
dateparser==1.1.8
decorator==5.2.1
dill==0.4.0
docopt==0.6.2
einops==0.8.2
einx==0.3.0
ema-pytorch==0.7.9
encodec==0.1.1
fastapi==0.115.6
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.8.0
gradio_client==2.2.0
groovy==0.1.2
grpcio==1.78.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonlines==1.2.0
kiwisolver==1.4.9
ko-speech-tools==0.1.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.4.2
omegaconf==2.3.0
openai-whisper==20250625
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==13.9.4
rjieba==0.2.0
rsa==4.9.1
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentry-sdk==2.53.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
starlette==0.41.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
tomli==2.4.0
tomlkit==0.13.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchcodec==0.10.0
torchdiffeq==0.2.5
tqdm==4.67.3
transformers==4.51.3
transformers-stream-generator==0.0.5
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.34.0
vocos==0.1.0
wandb==0.25.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
win32_setctime==1.2.0
x-transformers==2.16.1
xxhash==3.6.0
yarl==1.22.0
